Move pairs to Combinatorics, remove wrapping
