+++ Sources/Combinatorics/Pairs.swift
@@ -1,0 +1,14 @@
+//
+//  Pairs.swift
+//  Combinatorics
+//
+//  Created by James Bean on 8/28/17.
+//
+
+extension Sequence where SubSequence: Sequence {
+
+    /// - Returns: `Zip2Sequence` of 2-tuples comosed of adjacent values.
+    public var pairs: Zip2Sequence<Self,SubSequence> {
+        return zip(self,dropFirst())
+    }
+}
